rewritten the routes for super users, renamed the super user blade files and changed some locations
