+++ resources/views/posts/superConfirm.blade.php
@@ -6,9 +6,9 @@
     </div>
     {!! Form::model($post, [
         'method' => 'delete',
-        'route' => ['posts.destroy', $post->id]
+        'route' => ['superPosts.destroy', $post->id]
     ]) !!}
     {!! Form::submit('Ja ik ben zeker', ['class'=>'btn btn-danger']) !!}
-    <a href="{{ route('posts.index') }}" class="btn btn-primary">Annuleren!</a>
+    <a href="{{ route('superPosts.index') }}" class="btn btn-primary">Annuleren!</a>
     {!! Form::close() !!}
 @endsection
